// src/@codegen/kliquidity/types/rebalanceAutodriftStep.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { combineCodec, getEnumDecoder, getEnumEncoder, type FixedSizeCodec, type FixedSizeDecoder, type FixedSizeEncoder } from '@solana/kit';

export enum RebalanceAutodriftStep { Uninitialized, Autodrifting };

export type RebalanceAutodriftStepArgs = RebalanceAutodriftStep;

export function getRebalanceAutodriftStepEncoder(): FixedSizeEncoder<RebalanceAutodriftStepArgs> {
    return getEnumEncoder(RebalanceAutodriftStep);
}

export function getRebalanceAutodriftStepDecoder(): FixedSizeDecoder<RebalanceAutodriftStep> {
    return getEnumDecoder(RebalanceAutodriftStep);
}

export function getRebalanceAutodriftStepCodec(): FixedSizeCodec<RebalanceAutodriftStepArgs, RebalanceAutodriftStep> {
    return combineCodec(getRebalanceAutodriftStepEncoder(), getRebalanceAutodriftStepDecoder());
}