// src/@codegen/kliquidity/types/stakingRateSource.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { combineCodec, getEnumDecoder, getEnumEncoder, type FixedSizeCodec, type FixedSizeDecoder, type FixedSizeEncoder } from '@solana/kit';

export enum StakingRateSource { Constant, Scope };

export type StakingRateSourceArgs = StakingRateSource;

export function getStakingRateSourceEncoder(): FixedSizeEncoder<StakingRateSourceArgs> {
    return getEnumEncoder(StakingRateSource);
}

export function getStakingRateSourceDecoder(): FixedSizeDecoder<StakingRateSource> {
    return getEnumDecoder(StakingRateSource);
}

export function getStakingRateSourceCodec(): FixedSizeCodec<StakingRateSourceArgs, StakingRateSource> {
    return combineCodec(getStakingRateSourceEncoder(), getStakingRateSourceDecoder());
}